refactor: workflow approval
