+++ packages/plugin-workflow-approval/src/client/usage/h5/todos/TodosBlock.tsx
@@ -7,9 +7,9 @@ import { SearchBar, Tabs } from 'antd-mobile';
 import { collectionApprovalTodos } from '../../../common/collections/approvalRecords';
 import { collectionApprovals } from '../../../common/collections/approvals';
 import { collectionFlowNodes } from '../../../common/collections/flowNodes';
+import { todoCollection } from '../../../common/collections/userJobs';
+import { CollectionWorkflowNotice } from '../../../common/collections/workflowNotice';
 import { collectionWorkflows } from '../../../common/collections/workflows';
-import { CollectionWorkflowNotice } from '../collection/notice.collection';
-import { todoCollection } from '../collection/UserJobs.collection';
 import { useTranslation } from '../locale';
 import { TabDuplicateItem } from './component/TabDuplicateItem';
 import { TabPendingItem } from './component/TabPendingItem';
